Move colors and shapes into a common submenu
